Separate WakeWordListener into an interface and a base class
Also merged the functionality of ShortPhraseListener into WakeWordListenerBase.
