+++ Scripts/Dialog/IWakeWordListener.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 212c529aad5a7426cac598eb40d2f99f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
