+++ Scripts/Dialog/WakeWord.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2f85fdff6da10478dabbd7fe5c0cb0bb
+guid: bbe31019bdeb040568b2dbf032ef9921
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
