+++ Scripts/Dialog/WakeWordListenerBase.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Threading;
 using UnityEngine;
 using Cysharp.Threading.Tasks;
 using ChatdollKit.IO;
@@ -7,20 +8,45 @@ using ChatdollKit.Network;
 
 namespace ChatdollKit.Dialog
 {
-    public class WakeWordListenerBase : ShortPhraseListener
+    public class WakeWordListenerBase : VoiceRecorderBase, IWakeWordListener
     {
+        public bool AutoStart = true;
+        public Func<string, UniTask> OnRecognizedAsync;
+
+        [Header("Test and Debug")]
+        public bool PrintResult = false;
+
+        [Header("Voice Recorder Settings")]
+        public float VoiceDetectionThreshold = 0.1f;
+        public float VoiceDetectionRaisedThreshold = 0.5f;
+        public float VoiceDetectionMinimumLength = 0.2f;
+        public float SilenceDurationToEndRecording = 0.3f;
+        public float VoiceRecognitionMaximumLength = 3.0f;
+
+        public Action OnListeningStart;
+        public Action OnListeningStop;
+        public Action OnRecordingStart;
+        public Action<float> OnDetectVoice;
+        public Action<AudioClip> OnRecordingEnd;
+        public Action<Exception> OnError = (e) => { Debug.LogError($"Recording wakeword error: {e.Message}\n{e.StackTrace}"); };
+
+        // Protected members for recording voice and recognize task
+        protected CancellationTokenSource cancellationTokenSource;
+
         [Header("WakeWord Settings")]
         public List<WakeWord> WakeWords;
         public List<string> CancelWords;
         public List<string> IgnoreWords = new List<string>() { "。", "、", "？", "！" };
-        public Func<string, WakeWord> ExtractWakeWord;
-        public Func<string, string> ExtractCancelWord;
-        public Func<WakeWord, UniTask> OnWakeAsync;
-        public Func<UniTask> OnCancelAsync;
+
+        public Func<string, WakeWord> ExtractWakeWord { get; set; }
+        public Func<string, string> ExtractCancelWord { get; set; }
+        public Func<WakeWord, UniTask> OnWakeAsync { get; set; }
+        public Func<UniTask> OnCancelAsync { get; set; }
+        public Func<bool> ShouldRaiseThreshold { get; set; } = () => { return false; };
 
         protected ChatdollHttp client = new ChatdollHttp();
 
-        protected override void Start()
+        protected virtual void Start()
         {
             if (AutoStart)
             {
@@ -38,7 +64,96 @@ namespace ChatdollKit.Dialog
             }
         }
 
-        protected override async UniTask ProcessVoiceAsync(VoiceRecorderResponse voiceRecorderResponse)
+        protected virtual void Update()
+        {
+            // Observe which threshold should be applied in every frames
+            voiceDetectionThreshold = ShouldRaiseThreshold() ? VoiceDetectionRaisedThreshold : VoiceDetectionThreshold;
+        }
+
+        protected override void OnDestroy()
+        {
+            base.OnDestroy();
+            cancellationTokenSource?.Cancel();
+        }
+
+        public void SetWakeWord(WakeWord wakeWord)
+        {
+            foreach (var ww in WakeWords)
+            {
+                if (ww.Text == wakeWord.Text)
+                {
+                    return;
+                }
+            }
+
+            WakeWords.Add(wakeWord);
+        }
+
+        public void SetCancelWord(string cancelWord)
+        {
+            foreach (var cw in CancelWords)
+            {
+                if (cw == cancelWord)
+                {
+                    return;
+                }
+            }
+
+            CancelWords.Add(cancelWord);
+        }
+
+        protected virtual async UniTask StartListeningAsync()
+        {
+            if (IsListening)
+            {
+                Debug.LogWarning("WakeWordListener is already listening");
+                return;
+            }
+
+            StartListening();   // Start recorder here to asure that GetVoiceAsync will be called after recorder started
+
+            try
+            {
+                cancellationTokenSource = new CancellationTokenSource();
+                var token = cancellationTokenSource.Token;
+
+                while (!token.IsCancellationRequested)
+                {
+                    voiceDetectionThreshold = VoiceDetectionThreshold;
+                    voiceDetectionMinimumLength = VoiceDetectionMinimumLength;
+                    silenceDurationToEndRecording = SilenceDurationToEndRecording;
+                    onListeningStart = OnListeningStart;
+                    onListeningStop = OnListeningStop;
+                    onRecordingStart = OnRecordingStart;
+                    onDetectVoice = OnDetectVoice;
+                    onRecordingEnd = OnRecordingEnd;
+                    onError = OnError;
+
+                    var voiceRecorderResponse = await GetVoiceAsync(0.0f, token);
+                    if (voiceRecorderResponse != null)
+                    {
+                        if (!string.IsNullOrEmpty(voiceRecorderResponse.Text) ||
+                            (voiceRecorderResponse.Voice != null && voiceRecorderResponse.Voice.length <= VoiceRecognitionMaximumLength)
+                        )
+                        {
+#pragma warning disable CS4014
+                            ProcessVoiceAsync(voiceRecorderResponse); // Do not await to continue listening
+#pragma warning restore CS4014
+                        }
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                Debug.LogError($"Error occured in listening wakeword: {ex.Message}\n{ex.StackTrace}");
+            }
+            finally
+            {
+                StopListening();
+            }
+        }
+
+        protected async UniTask ProcessVoiceAsync(VoiceRecorderResponse voiceRecorderResponse)
         {
             var recognizedText = string.Empty;
             if (!string.IsNullOrEmpty(voiceRecorderResponse.Text))
@@ -90,6 +205,13 @@ namespace ChatdollKit.Dialog
             }
         }
 
+#pragma warning disable CS1998
+        protected virtual async UniTask<string> RecognizeSpeechAsync(VoiceRecorderResponse recordedVoice)
+        {
+            throw new NotImplementedException("RecognizeSpeechAsync method should be implemented at the sub class of WakeWordListenerBase");
+        }
+#pragma warning restore CS1998
+
         public WakeWord ExtractWakeWordDefault(string text)
         {
             var textLower = text.ToLower();
@@ -133,52 +255,6 @@ namespace ChatdollKit.Dialog
             }
 
             return string.Empty;
-        }
-
-#pragma warning disable CS1998
-        protected virtual async UniTask<string> RecognizeSpeechAsync(VoiceRecorderResponse recordedVoice)
-        {
-            throw new NotImplementedException("RecognizeSpeechAsync method should be implemented at the sub class of WakeWordListenerBase");
-        }
-#pragma warning restore CS1998
-    }
-
-    [Serializable]
-    public class WakeWord
-    {
-        public string Text;
-        public int PrefixAllowance = 4;
-        public int SuffixAllowance = 4;
-        public string Intent;
-        public Priority IntentPriority = Priority.Normal;
-        public RequestType RequestType = RequestType.Voice;
-        public int InlineRequestMinimumLength = 0;
-        public string RecognizedText { get; private set; }
-        public string InlineRequestText { get; private set; }
-        public bool SkipPrompt = false;
-
-        public WakeWord CloneWithRecognizedText(string recognizedText)
-        {
-            var ww = new WakeWord
-            {
-                Text = Text,
-                PrefixAllowance = PrefixAllowance,
-                SuffixAllowance = SuffixAllowance,
-                Intent = Intent,
-                IntentPriority = IntentPriority,
-                RequestType = RequestType,
-                InlineRequestMinimumLength = InlineRequestMinimumLength,
-                RecognizedText = recognizedText,
-                InlineRequestText = string.Empty,
-                SkipPrompt = SkipPrompt
-            };
-
-            if (SkipPrompt)
-            {
-                ww.InlineRequestText = ww.RecognizedText;
-            }
-
-            return ww;
         }
     }
 }
